refactor: consolidate compute services
