+++ verdesat/webapp/services/project_compute.py
@@ -2,11 +2,16 @@ from __future__ import annotations
 
 """Project-level computation of metrics for the web application."""
 
+from collections.abc import Iterator
+from contextlib import contextmanager
 from datetime import date
 import hashlib
 import json
 import logging
+import os
+import pickle
 import tempfile
+import io
 from concurrent.futures import ThreadPoolExecutor
 from pathlib import Path
 from typing import Dict, Tuple, Protocol, cast
@@ -16,6 +21,14 @@ import pandas as pd
 import streamlit as st
 from shapely.geometry import mapping
 
+try:  # pragma: no cover - optional dependency
+    import redis  # type: ignore
+except Exception:  # pragma: no cover - optional dependency
+    redis = None
+
+from verdesat.analytics.stats import compute_summary_stats
+from verdesat.analytics.timeseries import TimeSeries
+from verdesat.services.timeseries import download_timeseries
 from verdesat.biodiv.bscore import BScoreCalculator
 from verdesat.biodiv.metrics import MetricEngine
 from verdesat.services.msa import MSAService
@@ -25,7 +38,167 @@ from verdesat.core.config import ConfigManager
 from verdesat.core.logger import Logger
 from verdesat.core.storage import StorageAdapter
 
-from .compute import _load_cache, _persist_cache, _ndvi_stats, _msavi_stats
+CONFIG = ConfigManager(
+    str(Path(__file__).resolve().parents[2] / "resources" / "webapp.toml")
+)
+REDIS_URL = CONFIG.get("cache", {}).get("redis_url")
+logger = Logger.get_logger(__name__)
+
+
+def _cache_path(storage: StorageAdapter, key: str) -> str:
+    """Return path used for persisted caches."""
+
+    return storage.join("cache", f"{key}.pkl")
+
+
+def _df_to_bytes(df: pd.DataFrame) -> io.BytesIO:
+    """Return ``df`` as CSV in a buffer."""
+
+    buf = io.BytesIO()
+    df.to_csv(buf, index=False)
+    buf.seek(0)
+    return buf
+
+
+@contextmanager
+def _suppress_timeseries_logging() -> Iterator[None]:
+    """Silence timeseries logs to avoid Streamlit context errors."""
+
+    ts_logger = logging.getLogger("verdesat.services.timeseries")
+    prev = ts_logger.level
+    ts_logger.setLevel(logging.ERROR)
+    try:
+        yield
+    finally:
+        ts_logger.setLevel(prev)
+
+
+def _persist_cache(storage: StorageAdapter, key: str, value: object) -> None:
+    """Persist ``value`` under ``key`` using Redis or storage."""
+
+    data = pickle.dumps(value)
+    url = REDIS_URL
+    if redis and url:
+        try:  # pragma: no cover - network failure
+            redis.Redis.from_url(url).set(key, data)
+            logger.debug("stored cache %s in Redis", key)
+            return
+        except Exception:
+            logger.exception("failed to persist %s in Redis", key)
+    try:  # pragma: no cover - storage failure
+        storage.write_bytes(_cache_path(storage, key), data)
+        logger.debug("stored cache %s at %s", key, _cache_path(storage, key))
+    except Exception:
+        logger.exception("failed to persist %s to storage", key)
+
+
+def _load_cache(storage: StorageAdapter, key: str) -> object | None:
+    """Return persisted cache for ``key`` if available."""
+
+    url = REDIS_URL
+    if redis and url:
+        try:  # pragma: no cover - network failure
+            data = redis.Redis.from_url(url).get(key)
+            if data:
+                logger.debug("loaded cache %s from Redis", key)
+                return pickle.loads(data)
+        except Exception:
+            logger.exception("failed to load cache %s from Redis", key)
+    path = _cache_path(storage, key)
+    if os.path.exists(path):
+        try:  # pragma: no cover - I/O failure
+            with open(path, "rb") as fh:
+                logger.debug("loaded cache %s from %s", key, path)
+                return pickle.loads(fh.read())
+        except Exception:
+            logger.exception("failed to load cache %s from %s", key, path)
+    logger.debug("cache miss for %s", key)
+    return None
+
+
+def _stats_row_to_dict(row: pd.Series, index: str) -> dict[str, float | str]:
+    """Return flat mapping of metrics for vegetation *index*."""
+
+    label = index.upper()
+    return {
+        f"{index}_mean": float(row[f"Mean {label}"]),
+        f"{index}_median": float(row[f"Median {label}"]),
+        f"{index}_min": float(row[f"Min {label}"]),
+        f"{index}_max": float(row[f"Max {label}"]),
+        f"{index}_std": float(row[f"Std {label}"]),
+        f"{index}_slope": float(row[f"Sen's Slope ({label}/yr)"]),
+        f"{index}_delta": float(row[f"Trend Δ{label}"]),
+        f"{index}_p_value": float(row["Mann–Kendall p-value"]),
+        f"{index}_amp": float(row["Seasonal Amplitude"]),
+        f"{index}_resid_rms": float(row["Residual RMS"]),
+        f"{index}_peak": row["Peak Month"] if pd.notna(row["Peak Month"]) else "",
+        f"{index}_pct_fill": float(row["% Gapfilled"]),
+    }
+
+
+@st.cache_data
+def _ndvi_stats(
+    aoi_path: str, start_year: int, end_year: int
+) -> tuple[dict[str, float | str], pd.DataFrame]:
+    """Return NDVI stats and decomposition for ``aoi_path``."""
+
+    with _suppress_timeseries_logging():
+        ts_df = download_timeseries(
+            geojson=aoi_path,
+            collection="COPERNICUS/S2_SR_HARMONIZED",
+            start=f"{start_year}-01-01",
+            end=f"{end_year}-12-31",
+            index="ndvi",
+            chunk_freq="YE",
+            agg="ME",
+        )
+    ts = TimeSeries.from_dataframe(ts_df, index="ndvi").fill_gaps()
+    decomp = ts.decompose(period=12)
+    pid = ts.df["id"].iloc[0]
+    res = decomp.get(pid)
+    if res is None:  # pragma: no cover - requires non-empty series
+        raise ValueError("decomposition failed")
+    decomp_df = pd.DataFrame(
+        {
+            "date": res.observed.index,
+            "observed": res.observed.values,
+            "trend": res.trend.values,
+            "seasonal": res.seasonal.values,
+            "resid": res.resid.values,
+        }
+    )
+    ts_bytes = _df_to_bytes(ts.df)
+    decomp_bytes = _df_to_bytes(decomp_df)
+    stats_df = compute_summary_stats(
+        ts_bytes, decomp_dir={pid: decomp_bytes}, value_col="mean_ndvi"
+    ).to_dataframe()
+    row = stats_df.iloc[0]
+    stats = _stats_row_to_dict(row, "ndvi")
+    return stats, decomp_df[["date", "observed", "trend", "seasonal"]]
+
+
+@st.cache_data
+def _msavi_stats(
+    aoi_path: str, start_year: int, end_year: int
+) -> tuple[dict[str, float | str], pd.DataFrame]:
+    """Return MSAVI stats and monthly time series for ``aoi_path``."""
+
+    with _suppress_timeseries_logging():
+        ts_df = download_timeseries(
+            geojson=aoi_path,
+            collection="COPERNICUS/S2_SR_HARMONIZED",
+            start=f"{start_year}-01-01",
+            end=f"{end_year}-12-31",
+            index="msavi",
+            chunk_freq="YE",
+            agg="ME",
+        )
+    ts = TimeSeries.from_dataframe(ts_df, index="msavi").fill_gaps()
+    ts_bytes = _df_to_bytes(ts.df)
+    stats_df = compute_summary_stats(ts_bytes, value_col="mean_msavi").to_dataframe()
+    row = stats_df.iloc[0]
+    stats = _stats_row_to_dict(row, "msavi")
+    return stats, ts.df
 
 
 class ChipService(Protocol):
